+++ asilib/io/load.py
@@ -215,8 +215,8 @@ def load_skymap(
         time = dateutil.parser.parse(time)
 
     if force_download:
-        skymap_paths = download_skymap(asi_array_code.lower(),
-            location_code, force_download=force_download
+        skymap_paths = download_skymap(
+            asi_array_code.lower(), location_code, force_download=force_download
         )
 
     else:
@@ -233,8 +233,8 @@ def load_skymap(
 
         # Download skymap files if they are not downloaded yet.
         if len(skymap_paths) == 0:
-            skymap_paths = download_skymap(asi_array_code.lower(),
-                location_code, force_download=force_download
+            skymap_paths = download_skymap(
+                asi_array_code.lower(), location_code, force_download=force_download
             )
 
     skymap_dates = _extract_skymap_dates(skymap_paths)
@@ -414,7 +414,7 @@ def _load_images(
     ignore_missing_data: bool
         Flag to ignore the ``FileNotFoundError`` that is raised when ASI
         data is unavailable for that date-hour. Only useful when ``time_range``
-        is passed.
+        is passed. If no data is loaded at all, an AssertionError is raised.
 
     Returns
     -------
@@ -432,6 +432,9 @@ def _load_images(
         If the image dimensions are not specified for an ASI array.
     AssertionError
         If len(time_range) != 2.
+    AssertionError
+        If no time stamps were found in time_range, regardless of the
+        ignore_missing_data value.
 
     Example
     -------
@@ -460,9 +463,12 @@ def _load_images(
 
         start_time_index += file_images.shape[0]
 
-    i_nan = np.where(~np.isnan(images[:, 0, 0]))[0]
-    images = images[i_nan, :, :]
-    times = times[i_nan]
+    i_not_nan = np.where(~np.isnan(images[:, 0, 0]))[0]
+    assert len(i_not_nan) > 0, (
+        f'{len(i_not_nan)} number of time stamps' ' were found in time_range={time_range}'
+    )
+    images = images[i_not_nan, :, :]
+    times = times[i_not_nan]
     return times, images
 
 
@@ -515,8 +521,8 @@ def _find_img_path(
     time = utils._validate_time(time)
 
     if force_download:
-        file_path = download_image(asi_array_code.lower(), 
-            location_code, time=time, force_download=force_download
+        file_path = download_image(
+            asi_array_code.lower(), location_code, time=time, force_download=force_download
         )[0]
     else:
         # If the user does not want to force a download, look for a file on the
@@ -533,8 +539,8 @@ def _find_img_path(
 
         elif len(matched_paths) == 0:  # No local file found
             try:
-                file_path = download_image(asi_array_code.lower(),
-                    location_code, time=time, force_download=force_download
+                file_path = download_image(
+                    asi_array_code.lower(), location_code, time=time, force_download=force_download
                 )[0]
             except NotADirectoryError:
                 raise FileNotFoundError(