+++ asilib/tests/test_download_themis.py
@@ -42,8 +42,9 @@ class TestDownloadThemis(unittest.TestCase):
         temp_image_dir = pathlib.Path(asilib.config['ASI_DATA_DIR'], 'themis')
         temp_image_path = temp_image_dir / 'thg_l1_asf_gill_2016102904_v01.cdf'
 
-        save_path = asilib.download_image(self.asi_array_code, self.location_code, time=self.day, 
-            force_download=True)
+        save_path = asilib.download_image(
+            self.asi_array_code, self.location_code, time=self.day, force_download=True
+        )
 
         assert temp_image_path == save_path[0]
         assert temp_image_path.is_file()